+++ tinper-bee/bee-message/CHANGELOG.md
@@ -1,5 +1,15 @@
+<a name="2.0.3"></a>
+## [2.0.3](https://github.com/tinper-bee/bee-message/compare/v2.0.2...v2.0.3) (2019-05-07)
+
+
+
 <a name="2.0.2"></a>
-## [2.0.2](https://github.com/tinper-bee/bee-message/compare/v2.0.0...v2.0.2) (2019-04-24)
+## [2.0.2](https://github.com/tinper-bee/bee-message/compare/v2.0.1...v2.0.2) (2019-04-24)
+
+
+
+<a name="2.0.1"></a>
+## [2.0.1](https://github.com/tinper-bee/bee-message/compare/v2.0.0...v2.0.1) (2019-04-24)
 
 
 ### Bug Fixes